Rename file and update logic
